get rid of foast.Shift and ShiftRecognizer pass
