+++ .pre-commit-config.yaml
@@ -72,6 +72,7 @@ repos:
       src/eve/datamodels.py | # until datamodels is upgraded to 3.10
       src/functional/ffront/func_to_foast.py | # until mypy supports match
       src/functional/ffront/symbol_makers.py | # until mypy supports match
+      src/functional/ffront/type_info.py | # until mypy supports match
       src/functional/ffront/foast_passes/type_deduction.py | # until mypy supports match
       docs/eve/conf.py |
       tests/eve_tests/definitions.py |
