+++ tests/functional_tests/ffront_tests/test_type_deduction.py
@@ -18,8 +18,9 @@ from functional.common import Dimension
 from functional.ffront import common_types
 from functional.ffront import field_operator_ast as foast
 from functional.ffront.fbuiltins import Field, float64, int64
-from functional.ffront.foast_passes.type_deduction import FieldOperatorTypeDeductionError, TypeInfo
+from functional.ffront.foast_passes.type_deduction import FieldOperatorTypeDeductionError
 from functional.ffront.func_to_foast import FieldOperatorParser
+from functional.ffront.type_info import TypeInfo
 
 
 def type_info_cases():
